refactor: 컴포넌트 함수 분리
